Trying to use bezier curve instead of quadratic curve for parabola (fixes problem for OpenJDK 7).
git-svn-id: http://www.geogebra.org/svn/trunk/geogebra@15165 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ web/src/geogebra/web/awt/GeneralPath.java
@@ -134,8 +134,8 @@ public class GeneralPath extends geogebra.common.awt.GeneralPath implements
     }
 
 	@Override
-    public void curveTo(float f, float g, float h, float i, float j, float k) {
-		impl.curveTo(f, g, h, i, j, k);
+    public void curveTo(double f, double g, double h, double i, double j, double k) {
+		impl.curveTo((float)f, (float)g, (float)h, (float)i, (float)j, (float)k);
 	    
     }
